// frontend/src/types/elements.ts
export const enum ElementNames {
  TEXTBOX = 'textbox',
  TEXT = 'text',
  ITEXT = 'i-text',
  INPUTTEXT = 'itext',
  ARCTEXT = 'arctext',
  VERTICALTEXT = 'verticaltext',
  IMAGE = 'image',
  SVGIMAGE = 'svgimage',
  CROPIMAGE = 'cropimage',
  MASK = 'mask',
  PATH = 'path',
  RECT = 'rect',
  LINE = 'line',
  ARROW = 'arrow',
  POLYLINE = 'polyline',
  ELLIPSE = 'ellipse',
  QRCODE = 'qrcode',
  BARCODE = 'barcode',
  GROUP = 'group',
  ACTIVE = 'activeselection',
  CIRCLE = 'circle',
  REFERENCELINE = 'referenceline',
  DOT = 'dot',
}

export const SupportEffects = [
  'group',
  'activeselection',
  'itext',
  'text',
  'textbox',
  'image',
]

export interface ColorStop {
  color: string
  offset: number
  opacity?: number
}
 export interface PathGradient {
  type: 'linear' | 'radial'
  name: string
  color: string[]
  rotate: number
}


export type LinePoint = '' | 'arrow' | 'dot' 
// ...
export const enum RightStates {
  ELEMENT_CANVAS = 'design',
  ELEMENT_TEXT = 'text',
  ELEMENT_SVG = 'path',
  ELEMENT_IMAGE = 'image',
  ELEMENT_CODE = 'code',
  ELEMENT_STYLE = 'style',
  ELEMENT_POSITION = 'position',
  ELEMENT_LAYER = 'layer',
  ELEMENT_EFFECT = 'effect',
}